Got User Profile Working
Here is everything so far. The User Profile activity compiles but needs to add functionality to it to get user info. The Trail Map activity is not functional. I tried copying what I used for my individual project but it didn't quite work. It might be better to start over, but need input